pip-requirements.txt
## @file
# EDK II Python PIP requirements file
#
# This file provides the list of python components to install using PIP.
#
# Copyright (c) Microsoft Corporation.
# SPDX-License-Identifier: BSD-2-Clause-Patent
#
# https://pypi.org/project/pip/
# https://pip.pypa.io/en/stable/user_guide/#requirements-files
# https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format
# https://www.python.org/dev/peps/pep-0440/#version-specifiers
##

edk2-pytool-library==0.15.0
edk2-pytool-extensions==0.23.3
xmlschema==2.3.0
regex==2023.5.5
pyelftools=0.29
